+++ test/integration/handlers/transfers/handlers.test.js
@@ -29,44 +29,45 @@ const Test = require('tape')
 const Uuid = require('uuid4')
 const retry = require('async-retry')
 const Logger = require('@mojaloop/central-services-logger')
-const Config = require('../../../src/lib/config')
+const Config = require('#src/lib/config')
 const Time = require('@mojaloop/central-services-shared').Util.Time
 const sleep = Time.sleep
 const Db = require('@mojaloop/central-services-database').Db
-const Cache = require('../../../src/lib/cache')
+const Cache = require('#src/lib/cache')
 const Tb = require('../../../src/lib/tb')
 const Consumer = require('@mojaloop/central-services-stream').Util.Consumer
 const Producer = require('@mojaloop/central-services-stream').Util.Producer
 const Utility = require('@mojaloop/central-services-shared').Util.Kafka
 const Enum = require('@mojaloop/central-services-shared').Enum
-const ParticipantHelper = require('../helpers/participant')
-const ParticipantLimitHelper = require('../helpers/participantLimit')
-const ParticipantFundsInOutHelper = require('../helpers/participantFundsInOut')
-const ParticipantEndpointHelper = require('../helpers/participantEndpoint')
-const SettlementHelper = require('../helpers/settlementModels')
-const HubAccountsHelper = require('../helpers/hubAccounts')
-const TransferService = require('../../../src/domain/transfer')
-const ParticipantService = require('../../../src/domain/participant')
-const TransferExtensionModel = require('../../../src/models/transfer/transferExtension')
+const ParticipantHelper = require('#test/integration/helpers/participant')
+const ParticipantLimitHelper = require('#test/integration/helpers/participantLimit')
+const ParticipantFundsInOutHelper = require('#test/integration/helpers/participantFundsInOut')
+const ParticipantEndpointHelper = require('#test/integration/helpers/participantEndpoint')
+const SettlementHelper = require('#test/integration/helpers/settlementModels')
+const HubAccountsHelper = require('#test/integration/helpers/hubAccounts')
+const TransferService = require('#src/domain/transfer/index')
+const ParticipantService = require('#src/domain/participant/index')
+const TransferExtensionModel = require('#src/models/transfer/transferExtension')
 const Util = require('@mojaloop/central-services-shared').Util
 const ErrorHandler = require('@mojaloop/central-services-error-handling')
 const {
   wrapWithRetries,
   getMessagePayloadOrThrow,
   sleepPromise
-} = require('../../util/helpers')
-const TestConsumer = require('../helpers/testConsumer')
+} = require('#test/util/helpers')
+const TestConsumer = require('#test/integration/helpers/testConsumer')
+const KafkaHelper = require('#test/integration/helpers/kafkaHelper')
 
-const ParticipantCached = require('../../../src/models/participant/participantCached')
-const ParticipantCurrencyCached = require('../../../src/models/participant/participantCurrencyCached')
-const ParticipantLimitCached = require('../../../src/models/participant/participantLimitCached')
-const SettlementModelCached = require('../../../src/models/settlement/settlementModelCached')
+const ParticipantCached = require('#src/models/participant/participantCached')
+const ParticipantCurrencyCached = require('#src/models/participant/participantCurrencyCached')
+const ParticipantLimitCached = require('#src/models/participant/participantLimitCached')
+const SettlementModelCached = require('#src/models/settlement/settlementModelCached')
 
 const Handlers = {
-  index: require('../../../src/handlers/register'),
-  positions: require('../../../src/handlers/positions/handler'),
-  transfers: require('../../../src/handlers/transfers/handler'),
-  timeouts: require('../../../src/handlers/timeouts/handler')
+  index: require('#src/handlers/register'),
+  positions: require('#src/handlers/positions/handler'),
+  transfers: require('#src/handlers/transfers/handler'),
+  timeouts: require('#src/handlers/timeouts/handler')
 }
 
 const TransferState = Enum.Transfers.TransferState
@@ -74,10 +75,10 @@ const TransferInternalState = Enum.Transfers.TransferInternalState
 const TransferEventType = Enum.Events.Event.Type
 const TransferEventAction = Enum.Events.Event.Action
 
-const debug = false
-const rebalanceDelay = 10000
-const retryDelay = 500
-const retryCount = 40
+const debug = process?.env?.TEST_INT_DEBUG || false
+const rebalanceDelay = process?.env?.TEST_INT_REBALANCE_DELAY || 10000
+const retryDelay = process?.env?.TEST_INT_RETRY_DELAY || 2
+const retryCount = process?.env?.TEST_INT_RETRY_COUNT || 40
 const retryOpts = {
   retries: retryCount,
   minTimeout: retryDelay,
@@ -128,6 +129,38 @@ const testDataZAR = {
 
 const prepareTestData = async (dataObj) => {
   try {
+    // TODO: START - Disabling these handlers to test running the CL as a separate service independently.
+    //       The following issue https://github.com/mojaloop/project/issues/3112 was created to investigate as to why the Integration Tests are so unstable when then Event Handlers are executing in-line. For the time being the above PR clearly separates the process which resolves the stability issue for the time being.
+    // // Lets make sure that all existing producers are connected
+    // await KafkaHelper.producers.connect()
+    // // Lets make sure that all existing Consumers are connected
+    // await KafkaHelper.consumers.connect()
+    // const topics = TestTopics.list
+    // TODO: END - Disabling these handlers to test running the CL as a separate service independently.
+
+    // // lets make sure all our Producers are already connected if they have already been defined.
+    // for (const topic of topics) {
+    //   try {
+    //     // lets make sure check if any of our Producers are already connected if they have already been defined.
+    //     console.log(`Producer[${topic}] checking connectivity!`)
+    //     const isConnected = await Producer.isConnected(topic)
+    //     if (!isConnected) {
+    //       try {
+    //         console.log(`Producer[${topic}] is connecting`)
+    //         await Producer.getProducer(topic).connect()
+    //         console.log(`Producer[${topic}] is connected`)
+    //       } catch (err) {
+    //         console.log(`Producer[${topic}] connection failed!`)
+    //       }
+    //     } else {
+    //       console.log(`Producer[${topic}] is ALREADY connected`)
+    //     }
+    //   } catch (err) {
+    //     console.log(`Producer[${topic}] has not been initialized`)
+    //     console.error(err)
+    //   }
+    // }
+
     const payer = await ParticipantHelper.prepareData(dataObj.payer.name, dataObj.amount.currency)
     const payee = await ParticipantHelper.prepareData(dataObj.payee.name, dataObj.amount.currency)
 
@@ -297,6 +330,11 @@ Test('Handlers test', async handlersTest => {
   await SettlementHelper.prepareData()
   await HubAccountsHelper.prepareData()
 
+  const wrapWithRetriesConf = {
+    remainingRetries: retryOpts?.retries || 10, // default 10
+    timeout: retryOpts?.maxTimeout || 2 // default 2
+  }
+
   // Start a testConsumer to monitor events that our handlers emit
   const testConsumer = new TestConsumer([
     {
@@ -329,14 +367,18 @@ Test('Handlers test', async handlersTest => {
 
   await handlersTest.test('registerAllHandlers should', async registerAllHandlers => {
     await registerAllHandlers.test('setup handlers', async (test) => {
-      await Handlers.transfers.registerPrepareHandler()
-      await Handlers.positions.registerPositionHandler()
-      await Handlers.transfers.registerFulfilHandler()
-      await Handlers.timeouts.registerTimeoutHandler()
+      // TODO: START - Disabling these handlers to test running the CL as a separate service independently.
+      //       The following issue https://github.com/mojaloop/project/issues/3112 was created to investigate as to why the Integration Tests are so unstable when then Event Handlers are executing in-line. For the time being the above PR clearly separates the process which resolves the stability issue for the time being.
+      // await Handlers.transfers.registerPrepareHandler()
+      // await Handlers.positions.registerPositionHandler()
+      // await Handlers.transfers.registerFulfilHandler()
+      // await Handlers.timeouts.registerTimeoutHandler()
+      // TODO: END - Disabling these handlers to test running the CL as a separate service independently.
 
       // Set up the testConsumer here
       await testConsumer.startListening()
 
+      // TODO: MIG - Disabling these handlers to test running the CL as a separate service independently.
       sleep(rebalanceDelay, debug, 'registerAllHandlers', 'awaiting registration of common handlers')
 
       test.pass('done')
@@ -359,16 +401,24 @@ Test('Handlers test', async handlersTest => {
         TransferEventType.PREPARE.toUpperCase())
       prepareConfig.logger = Logger
       await Producer.produceMessage(td.messageProtocolPrepare, td.topicConfTransferPrepare, prepareConfig)
-      const transfer = await wrapWithRetries(async () => {
-        // lets fetch the transfer
-        const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-        // lets check its status, and if its what we expect return the result
-        if (transfer.transferState === 'RESERVED') return transfer
-        // otherwise lets return nothing
-        return null
-      })
 
-      test.equal(transfer.transferState, 'RESERVED', 'Transfer is in reserved state')
+      let transfer = {}
+      try {
+        transfer = await wrapWithRetries(async () => {
+          // lets fetch the transfer
+          const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
+          console.dir(transfer)
+          // lets check its status, and if its what we expect return the result
+          if (transfer?.transferState === 'RESERVED') return transfer
+          // otherwise lets return nothing
+          return null
+        }, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
+      }
+
+      test.equal(transfer?.transferState, 'RESERVED', 'Transfer is in reserved state')
 
       // 2. send an ABORTED request from Payee
       td.messageProtocolFulfil.metadata.event.action = TransferEventAction.RESERVE
@@ -392,13 +442,13 @@ Test('Handlers test', async handlersTest => {
         await wrapWithRetries(() => testConsumer.getEventsForFilter({
           topicFilter: 'topic-notification-event',
           action: 'reserved-aborted'
-        }))
+        }), wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
         test.notOk('Should not be executed')
       } catch (err) {
         console.log(err)
         test.ok('No payee abort notification sent')
       }
-      console.log(JSON.stringify(testConsumer.getAllEvents()))
+      if (debug) console.log(JSON.stringify(testConsumer.getAllEvents()))
 
       // TODO: I can't seem to find the payer abort notification in the log
       // is there something I'm missing here? Does it go to a different handler?
@@ -431,18 +481,26 @@ Test('Handlers test', async handlersTest => {
         TransferEventType.PREPARE.toUpperCase())
       prepareConfig.logger = Logger
       await Producer.produceMessage(td.messageProtocolPrepare, td.topicConfTransferPrepare, prepareConfig)
-      const transfer = await wrapWithRetries(async () => {
-        // lets fetch the transfer
-        const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-        // lets check its status, and if its what we expect return the result
-        if (transfer.transferState === 'RESERVED') return transfer
-        // otherwise lets return nothing
-        return null
-      })
-      test.equal(transfer.transferState, 'RESERVED', 'Transfer is in reserved state')
+
+      let transfer = {}
+      try {
+        transfer = await wrapWithRetries(async () => {
+          // lets fetch the transfer
+          const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
+          // lets check its status, and if its what we expect return the result
+          if (transfer?.transferState === 'RESERVED') return transfer
+          // otherwise lets return nothing
+          return null
+        }, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
+      }
+
+      test.equal(transfer?.transferState, 'RESERVED', 'Transfer is in reserved state')
 
       // 2. sleep so that the RESERVED transfer expires
-      await sleepPromise(2)
+      await sleepPromise(wrapWithRetriesConf.timeout)
 
       // 3. send a RESERVED request from Payee
       td.messageProtocolFulfil.metadata.event.action = TransferEventAction.RESERVE
@@ -462,27 +520,46 @@ Test('Handlers test', async handlersTest => {
 
       // 4. Get the updated transfer since the completedTimestamp may have changed
       const updatedTransfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-      const expectedAbortNotificationPayload = {
-        completedTimestamp: Time.getUTCString(new Date(updatedTransfer.completedTimestamp)),
-        transferState: 'ABORTED'
+
+      let expectedAbortNotificationPayload = {}
+      if (updatedTransfer) {
+        expectedAbortNotificationPayload = {
+          completedTimestamp: Time.getUTCString(new Date(updatedTransfer.completedTimestamp)),
+          transferState: 'ABORTED'
+        }
       }
 
       // Assert
       // 5. Check that we sent 2 notifications to kafka - one for the Payee, one for the Payer
-      const payerAbortNotification = (await wrapWithRetries(
-        () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'commit' }))
-      )[0]
-      const payeeAbortNotification = (await wrapWithRetries(
-        () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'reserved-aborted' }))
-      )[0]
+      let payerAbortNotification
+      let payeeAbortNotification
+      try {
+        payerAbortNotification = (await wrapWithRetries(
+          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'commit' }),
+          wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+        )[0]
+        payeeAbortNotification = (await wrapWithRetries(
+          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'reserved-aborted' }),
+          wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+        )[0]
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
+      }
+
       test.ok(payerAbortNotification, 'Payer Abort notification sent')
       test.ok(payeeAbortNotification, 'Payee Abort notification sent')
 
-      test.deepEqual(
-        getMessagePayloadOrThrow(payeeAbortNotification),
-        expectedAbortNotificationPayload,
-        'Abort notification should be sent with the correct values'
-      )
+      try {
+        test.deepEqual(
+          getMessagePayloadOrThrow(payeeAbortNotification),
+          expectedAbortNotificationPayload,
+          'Abort notification should be sent with the correct values'
+        )
+      } catch (err) {
+        test.notOk('Error should not be thrown - getMessagePayloadOrThrow(payeeAbortNotification) failed!')
+        console.error(err)
+      }
 
       // Cleanup
       testConsumer.clearEvents()
@@ -501,15 +578,23 @@ Test('Handlers test', async handlersTest => {
         TransferEventType.PREPARE.toUpperCase())
       prepareConfig.logger = Logger
       await Producer.produceMessage(td.messageProtocolPrepare, td.topicConfTransferPrepare, prepareConfig)
-      const transfer = await wrapWithRetries(async () => {
-        // lets fetch the transfer
-        const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-        // lets check its status, and if its what we expect return the result
-        if (transfer.transferState === 'RESERVED') return transfer
-        // otherwise lets return nothing
-        return null
-      })
-      test.equal(transfer.transferState, 'RESERVED', 'Transfer is in reserved state')
+
+      let transfer = {}
+      try {
+        transfer = await wrapWithRetries(async () => {
+          // lets fetch the transfer
+          const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
+          // lets check its status, and if its what we expect return the result
+          if (transfer?.transferState === 'RESERVED') return transfer
+          // otherwise lets return nothing
+          return null
+        }, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
+      }
+
+      test.equal(transfer?.transferState, 'RESERVED', 'Transfer is in reserved state')
 
       // 2. Modify the transfer in the DB
       await TransferService.saveTransferStateChange({
@@ -535,16 +620,21 @@ Test('Handlers test', async handlersTest => {
 
       // 4. Get the updated transfer since the completedTimestamp may have changed
       const updatedTransfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-      const expectedAbortNotificationPayload = {
-        completedTimestamp: Time.getUTCString(new Date(updatedTransfer.completedTimestamp)),
-        transferState: 'ABORTED'
+
+      let expectedAbortNotificationPayload = {}
+      if (updatedTransfer) {
+        expectedAbortNotificationPayload = {
+          completedTimestamp: Time.getUTCString(new Date(updatedTransfer.completedTimestamp)),
+          transferState: 'ABORTED'
+        }
       }
 
       // Assert
       // 5. Check that we sent 2 notifications to kafka - one for the Payee, one for the Payer
       try {
         const payerAbortNotification = (await wrapWithRetries(
-          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'commit' }))
+          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'commit' }),
+          wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
         )[0]
         test.ok(payerAbortNotification, 'Payer Abort notification sent')
       } catch (err) {
@@ -552,7 +642,8 @@ Test('Handlers test', async handlersTest => {
       }
       try {
         const payeeAbortNotification = (await wrapWithRetries(
-          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'reserved-aborted' }))
+          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'reserved-aborted' }),
+          wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
         )[0]
         test.ok(payeeAbortNotification, 'Payee Abort notification sent')
         test.deepEqual(
@@ -581,15 +672,23 @@ Test('Handlers test', async handlersTest => {
         TransferEventType.PREPARE.toUpperCase())
       prepareConfig.logger = Logger
       await Producer.produceMessage(td.messageProtocolPrepare, td.topicConfTransferPrepare, prepareConfig)
-      const transfer = await wrapWithRetries(async () => {
-        // lets fetch the transfer
-        const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-        // lets check its status, and if its what we expect return the result
-        if (transfer.transferState === 'RESERVED') return transfer
-        // otherwise lets return nothing
-        return null
-      })
-      test.equal(transfer.transferState, 'RESERVED', 'Transfer is in reserved state')
+
+      let transfer = {}
+      try {
+        transfer = await wrapWithRetries(async () => {
+          // lets fetch the transfer
+          const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
+          // lets check its status, and if its what we expect return the result
+          if (transfer?.transferState === 'RESERVED') return transfer
+          // otherwise lets return nothing
+          return null
+        }, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
+      }
+
+      test.equal(transfer?.transferState, 'RESERVED', 'Transfer is in reserved state')
 
       // 2. send a RESERVED request with an invalid validation(from Payee)
       td.messageProtocolFulfil.metadata.event.action = TransferEventAction.RESERVE
@@ -607,25 +706,45 @@ Test('Handlers test', async handlersTest => {
       fulfilConfig.logger = Logger
       await Producer.produceMessage(td.messageProtocolFulfil, td.topicConfTransferFulfil, fulfilConfig)
 
-      await wrapWithRetries(async () => {
-        const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-        return transfer.transferState === 'ABORTED_ERROR'
-      })
-      const updatedTransfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
-      test.equal(updatedTransfer.transferState, 'ABORTED_ERROR', 'Transfer is in ABORTED_ERROR state')
-      const expectedAbortNotificationPayload = {
-        completedTimestamp: (new Date(Date.parse(updatedTransfer.completedTimestamp))).toISOString(),
-        transferState: 'ABORTED'
+      try {
+        await wrapWithRetries(async () => {
+          const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
+          return transfer?.transferState === 'ABORTED_ERROR'
+        }, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
       }
 
-      // Assert
-      // 3. Check that we sent 2 notifications to kafka - one for the Payee, one for the Payer
-      const payerAbortNotificationEvent = (await wrapWithRetries(
-        () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'abort-validation' }))
-      )[0]
-      const payeeAbortNotificationEvent = (await wrapWithRetries(
-        () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'reserved-aborted' }))
-      )[0]
+      const updatedTransfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId)
+      test.equal(updatedTransfer?.transferState, 'ABORTED_ERROR', 'Transfer is in ABORTED_ERROR state')
+
+      let expectedAbortNotificationPayload = {}
+      if (updatedTransfer) {
+        expectedAbortNotificationPayload = {
+          completedTimestamp: (new Date(Date.parse(updatedTransfer.completedTimestamp))).toISOString(),
+          transferState: 'ABORTED'
+        }
+      }
+
+      let payerAbortNotificationEvent
+      let payeeAbortNotificationEvent
+      try {
+        // Assert
+        // 3. Check that we sent 2 notifications to kafka - one for the Payee, one for the Payer
+        payerAbortNotificationEvent = (await wrapWithRetries(
+          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'abort-validation' }),
+          wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+        )[0]
+        payeeAbortNotificationEvent = (await wrapWithRetries(
+          () => testConsumer.getEventsForFilter({ topicFilter: 'topic-notification-event', action: 'reserved-aborted' }),
+          wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+        )[0]
+      } catch (err) {
+        test.notOk('Error should not be thrown')
+        console.error(err)
+      }
+
       test.ok(payerAbortNotificationEvent, 'Payer Abort notification sent')
       test.ok(payeeAbortNotificationEvent, 'Payee Abort notification sent')
 
@@ -672,16 +791,16 @@ Test('Handlers test', async handlersTest => {
         const payerExpectedPosition = payerInitialPosition + td.transferPayload.amount.amount
         const payerPositionChange = await ParticipantService.getPositionChangeByParticipantPositionId(payerCurrentPosition.participantPositionId) || {}
         test.equal(producerResponse, true, 'Producer for prepare published message')
-        test.equal(transfer.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
+        test.equal(transfer?.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
         test.equal(payerCurrentPosition.value, payerExpectedPosition, 'Payer position incremented by transfer amount and updated in participantPosition')
         test.equal(payerPositionChange.value, payerCurrentPosition.value, 'Payer position change value inserted and matches the updated participantPosition value')
-        test.equal(payerPositionChange.transferStateChangeId, transfer.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
+        test.equal(payerPositionChange.transferStateChangeId, transfer?.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.RESERVED) {
+          if (transfer?.transferState !== TransferState.RESERVED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#1 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -711,17 +830,17 @@ Test('Handlers test', async handlersTest => {
         const payeeExpectedPosition = payeeInitialPosition - td.transferPayload.amount.amount
         const payeePositionChange = await ParticipantService.getPositionChangeByParticipantPositionId(payeeCurrentPosition.participantPositionId) || {}
         test.equal(producerResponse, true, 'Producer for fulfil published message')
-        test.equal(transfer.transferState, TransferState.COMMITTED, `Transfer state changed to ${TransferState.COMMITTED}`)
+        test.equal(transfer?.transferState, TransferState.COMMITTED, `Transfer state changed to ${TransferState.COMMITTED}`)
         test.equal(transfer.fulfilment, td.fulfilPayload.fulfilment, 'Commit ilpFulfilment saved')
         test.equal(payeeCurrentPosition.value, payeeExpectedPosition, 'Payee position decremented by transfer amount and updated in participantPosition')
         test.equal(payeePositionChange.value, payeeCurrentPosition.value, 'Payee position change value inserted and matches the updated participantPosition value')
-        test.equal(payeePositionChange.transferStateChangeId, transfer.transferStateChangeId, 'Payee position change record is bound to the corresponding transfer state change')
+        test.equal(payeePositionChange.transferStateChangeId, transfer?.transferStateChangeId, 'Payee position change record is bound to the corresponding transfer state change')
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.COMMITTED) {
+          if (transfer?.transferState !== TransferState.COMMITTED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#2 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -756,16 +875,16 @@ Test('Handlers test', async handlersTest => {
         const payerExpectedPosition = payerInitialPosition + td.transferPayload.amount.amount
         const payerPositionChange = await ParticipantService.getPositionChangeByParticipantPositionId(payerCurrentPosition.participantPositionId) || {}
         test.equal(producerResponse, true, 'Producer for prepare published message')
-        test.equal(transfer.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
+        test.equal(transfer?.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
         test.equal(payerCurrentPosition.value, payerExpectedPosition, 'Payer position incremented by transfer amount and updated in participantPosition')
         test.equal(payerPositionChange.value, payerCurrentPosition.value, 'Payer position change value inserted and matches the updated participantPosition value')
-        test.equal(payerPositionChange.transferStateChangeId, transfer.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
+        test.equal(payerPositionChange.transferStateChangeId, transfer?.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.RESERVED) {
+          if (transfer?.transferState !== TransferState.RESERVED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#1 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -793,17 +912,17 @@ Test('Handlers test', async handlersTest => {
         const payeeExpectedPosition = payeeInitialPosition - td.transferPayload.amount.amount
         const payeePositionChange = await ParticipantService.getPositionChangeByParticipantPositionId(payeeCurrentPosition.participantPositionId) || {}
         test.equal(producerResponse, true, 'Producer for fulfil published message')
-        test.equal(transfer.transferState, TransferState.COMMITTED, `Transfer state changed to ${TransferState.COMMITTED}`)
+        test.equal(transfer?.transferState, TransferState.COMMITTED, `Transfer state changed to ${TransferState.COMMITTED}`)
         test.equal(transfer.fulfilment, td.fulfilPayload.fulfilment, 'Commit ilpFulfilment saved')
         test.equal(payeeCurrentPosition.value, payeeExpectedPosition, 'Payee position decremented by transfer amount and updated in participantPosition')
         test.equal(payeePositionChange.value, payeeCurrentPosition.value, 'Payee position change value inserted and matches the updated participantPosition value')
-        test.equal(payeePositionChange.transferStateChangeId, transfer.transferStateChangeId, 'Payee position change record is bound to the corresponding transfer state change')
+        test.equal(payeePositionChange.transferStateChangeId, transfer?.transferStateChangeId, 'Payee position change record is bound to the corresponding transfer state change')
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.COMMITTED) {
+          if (transfer?.transferState !== TransferState.COMMITTED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#2 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -835,13 +954,13 @@ Test('Handlers test', async handlersTest => {
       const tests = async () => {
         const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
         test.equal(producerResponse, true, 'Producer for prepare published message')
-        test.equal(transfer.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
+        test.equal(transfer?.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.RESERVED) {
+          if (transfer?.transferState !== TransferState.RESERVED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#3 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -872,13 +991,13 @@ Test('Handlers test', async handlersTest => {
       const tests = async () => {
         const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
         test.equal(producerResponse, true, 'Producer for prepare published message')
-        test.equal(transfer.transferState, TransferInternalState.ABORTED_REJECTED, `Transfer state changed to ${TransferInternalState.ABORTED_REJECTED}`)
+        test.equal(transfer?.transferState, TransferInternalState.ABORTED_REJECTED, `Transfer state changed to ${TransferInternalState.ABORTED_REJECTED}`)
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferInternalState.ABORTED_REJECTED) {
+          if (transfer?.transferState !== TransferInternalState.ABORTED_REJECTED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#4 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -911,13 +1030,13 @@ Test('Handlers test', async handlersTest => {
       const tests = async () => {
         const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
         test.equal(producerResponse, true, 'Producer for prepare published message')
-        test.equal(transfer.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
+        test.equal(transfer?.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
       }
 
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.RESERVED) {
+          if (transfer?.transferState !== TransferState.RESERVED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#5 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -949,14 +1068,14 @@ Test('Handlers test', async handlersTest => {
         const transferError = await TransferService.getTransferErrorByTransferId(transfer.transferId)
         const transferExtension = await TransferExtensionModel.getByTransferId(transfer.transferId, false, true)
         test.equal(producerResponse, true, 'Producer for fulfil published message')
-        test.equal(transfer.transferState, TransferInternalState.ABORTED_ERROR, `Transfer state changed to ${TransferInternalState.ABORTED_ERROR}`)
+        test.equal(transfer?.transferState, TransferInternalState.ABORTED_ERROR, `Transfer state changed to ${TransferInternalState.ABORTED_ERROR}`)
         test.equal(payerCurrentPosition.value, payerExpectedPosition, 'Payer position decremented by transfer amount and updated in participantPosition')
         test.equal(payerPositionChange.value, payerCurrentPosition.value, 'Payer position change value inserted and matches the updated participantPosition value')
-        test.equal(payerPositionChange.transferStateChangeId, transfer.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
+        test.equal(payerPositionChange.transferStateChangeId, transfer?.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
         test.ok(transferError, 'A transfer error has been recorded')
         test.equal(transferError.errorCode, td.errorPayload.errorInformation.errorCode, 'Transfer error code matches')
         test.equal(transferError.errorDescription, expectedErrorDescription, 'Transfer error description matches')
-        test.notEqual(transferError.transferStateChangeId, transfer.transferStateChangeId, 'Transfer error record is bound to previous state of transfer')
+        test.notEqual(transferError.transferStateChangeId, transfer?.transferStateChangeId, 'Transfer error record is bound to previous state of transfer')
         test.ok(transferExtension, 'A transfer extension has been recorded')
         test.equal(transferExtension[0].transferId, transfer.transferId, 'Transfer extension recorded with transferErrorId key')
       }
@@ -964,7 +1083,7 @@ Test('Handlers test', async handlersTest => {
       try {
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferInternalState.ABORTED_ERROR) {
+          if (transfer?.transferState !== TransferInternalState.ABORTED_ERROR) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw ErrorHandler.Factory.createFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes.INTERNAL_SERVER_ERROR, `#6 Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
@@ -993,6 +1112,7 @@ Test('Handlers test', async handlersTest => {
       config.logger = Logger
 
       const producerResponse = await Producer.produceMessage(td.messageProtocolPrepare, td.topicConfTransferPrepare, config)
+      Logger.info(producerResponse)
 
       const tests = async () => {
         const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
@@ -1001,27 +1121,90 @@ Test('Handlers test', async handlersTest => {
         const payerExpectedPosition = payerInitialPosition + td.transferPayload.amount.amount
         const payerPositionChange = await ParticipantService.getPositionChangeByParticipantPositionId(payerCurrentPosition.participantPositionId) || {}
         test.equal(producerResponse, true, 'Producer for prepare published message')
-        test.equal(transfer.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
+        test.equal(transfer?.transferState, TransferState.RESERVED, `Transfer state changed to ${TransferState.RESERVED}`)
         test.equal(payerCurrentPosition.value, payerExpectedPosition, 'Payer position incremented by transfer amount and updated in participantPosition')
         test.equal(payerPositionChange.value, payerCurrentPosition.value, 'Payer position change value inserted and matches the updated participantPosition value')
-        test.equal(payerPositionChange.transferStateChangeId, transfer.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
+        test.equal(payerPositionChange.transferStateChangeId, transfer?.transferStateChangeId, 'Payer position change record is bound to the corresponding transfer state change')
       }
 
       try {
+        const retryTimeoutOpts = {
+          retries: Number(retryOpts.retries) * 2,
+          minTimeout: retryOpts.minTimeout,
+          maxTimeout: retryOpts.maxTimeout
+        }
+
         await retry(async () => { // use bail(new Error('to break before max retries'))
           const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
-          if (transfer.transferState !== TransferState.RESERVED) {
+          if (transfer?.transferState !== TransferState.RESERVED) {
             if (debug) console.log(`retrying in ${retryDelay / 1000}s..`)
             throw new Error(`#7   Max retry count ${retryCount} reached after ${retryCount * retryDelay / 1000}s. Tests fail`)
           }
           return tests()
-        }, retryOpts)
+        }, retryTimeoutOpts)
       } catch (err) {
         Logger.error(err)
         test.fail(err.message)
       }
 
       test.end()
+    })
+
+    await timeoutTest.test('update transfer after timeout with timeout status & error', async (test) => {
+      // Arrange
+      // Nothing to do here...
+
+      // Act
+
+      // Re-try function with conditions
+      const inspectTransferState = async () => {
+        try {
+          // Fetch Transfer record
+          const transfer = await TransferService.getById(td.messageProtocolPrepare.content.payload.transferId) || {}
+
+          // Check Transfer for correct state
+          if (transfer?.transferState === Enum.Transfers.TransferInternalState.EXPIRED_RESERVED) {
+            // We have a Transfer with the correct state, lets check if we can get the TransferError record
+            try {
+              // Fetch the TransferError record
+              const transferError = await TransferService.getTransferErrorByTransferId(td.messageProtocolPrepare.content.payload.transferId)
+              // TransferError record found, so lets return it
+              return {
+                transfer,
+                transferError
+              }
+            } catch (err) {
+              // NO TransferError record found, so lets return the transfer and the error
+              return {
+                transfer,
+                err
+              }
+            }
+          } else {
+            // NO Transfer with the correct state was found, so we return false
+            return false
+          }
+        } catch (err) {
+          // NO Transfer with the correct state was found, so we return false
+          Logger.error(err)
+          return false
+        }
+      }
+
+      // wait until we inspect a transfer with the correct status, or return false if all re-try attempts have failed
+      const result = await wrapWithRetries(inspectTransferState, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
+
+      // Assert
+      if (result === false) {
+        test.fail(`Transfer['${td.messageProtocolPrepare.content.payload.transferId}'].TransferState failed to transition to ${Enum.Transfers.TransferInternalState.EXPIRED_RESERVED}`)
+        test.end()
+      } else {
+        test.equal(result.transfer && result.transfer?.transferState, Enum.Transfers.TransferInternalState.EXPIRED_RESERVED, `Transfer['${td.messageProtocolPrepare.content.payload.transferId}'].TransferState = ${Enum.Transfers.TransferInternalState.EXPIRED_RESERVED}`)
+        test.equal(result.transferError && result.transferError.errorCode, ErrorHandler.Enums.FSPIOPErrorCodes.TRANSFER_EXPIRED.code, `Transfer['${td.messageProtocolPrepare.content.payload.transferId}'].transferError.errorCode = ${ErrorHandler.Enums.FSPIOPErrorCodes.TRANSFER_EXPIRED.code}`)
+        test.equal(result.transferError && result.transferError.errorDescription, ErrorHandler.Enums.FSPIOPErrorCodes.TRANSFER_EXPIRED.message, `Transfer['${td.messageProtocolPrepare.content.payload.transferId}'].transferError.errorDescription = ${ErrorHandler.Enums.FSPIOPErrorCodes.TRANSFER_EXPIRED.message}`)
+        test.pass()
+        test.end()
+      }
     })
 
     await timeoutTest.test('position resets after a timeout', async (test) => {
@@ -1034,7 +1217,7 @@ Test('Handlers test', async handlersTest => {
         return payerCurrentPosition.value === payerInitialPosition
       }
       // wait until we know the position reset, or throw after 5 tries
-      await wrapWithRetries(payerPositionDidReset, 10, 4)
+      await wrapWithRetries(payerPositionDidReset, wrapWithRetriesConf.remainingRetries, wrapWithRetriesConf.timeout)
       const payerCurrentPosition = await ParticipantService.getPositionByParticipantCurrencyId(td.payer.participantCurrencyId) || {}
 
       // Assert
@@ -1055,28 +1238,30 @@ Test('Handlers test', async handlersTest => {
       assert.pass('database connection closed')
       await testConsumer.destroy()
 
-      const topics = [
-        'topic-transfer-prepare',
-        'topic-transfer-position',
-        'topic-transfer-fulfil',
-        'topic-notification-event'
-      ]
-      for (const topic of topics) {
-        try {
-          await Producer.getProducer(topic).disconnect()
-          assert.pass(`producer to ${topic} disconnected`)
-        } catch (err) {
-          assert.pass(err.message)
-        }
-      }
-      for (const topic of topics) {
-        try {
-          await Consumer.getConsumer(topic).disconnect()
-          assert.pass(`consumer to ${topic} disconnected`)
-        } catch (err) {
-          assert.pass(err.message)
-        }
-      }
+      // TODO: Story to investigate as to why the Producers failed reconnection on the ./transfers/handlers.test.js - https://github.com/mojaloop/project/issues/3067
+      // const topics = KafkaHelper.topics
+      // for (const topic of topics) {
+      //   try {
+      //     await Producer.getProducer(topic).disconnect()
+      //     assert.pass(`producer to ${topic} disconnected`)
+      //   } catch (err) {
+      //     assert.pass(err.message)
+      //   }
+      // }
+      // Lets make sure that all existing Producers are disconnected
+      await KafkaHelper.producers.disconnect()
+
+      // TODO: Clean this up once the above issue has been resolved.
+      // for (const topic of topics) {
+      //   try {
+      //     await Consumer.getConsumer(topic).disconnect()
+      //     assert.pass(`consumer to ${topic} disconnected`)
+      //   } catch (err) {
+      //     assert.pass(err.message)
+      //   }
+      // }
+      // Lets make sure that all existing Consumers are disconnected
+      await KafkaHelper.consumers.disconnect()
 
       if (debug) {
         const elapsedTime = Math.round(((new Date()) - startTime) / 100) / 10